+++ src/main/java/com/vbutrim/weather/WeatherForecastManager.java
@@ -3,5 +3,5 @@ package com.vbutrim.weather;
 public interface WeatherForecastManager {
 
 
-    WeatherResponse getWeatherForecastByCityId(long cityId);
+    WeatherResponse getWeatherForecastByCityId(int cityId);
 }
